make Query a final class
